+++ web/app/page.tsx
@@ -1,0 +1,5 @@
+import { CategoryFinder } from "./category-finder";
+
+export default function Home() {
+  return <CategoryFinder />;
+}